requirements.txt
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-appconf==1.0.2
django-compressor==2.2
django-debug-permissions==1.0.0
django-debug-toolbar==1.11
django-extensions==2.1.4
django-haystack==2.8.1
django-libsass==0.7
django-markdownx==2.0.28
django-taggit==0.23.0
djangorestframework==3.9.1
idna==2.8
itypes==1.1.0
Jinja2==2.10
libsass==0.17.0
Markdown==3.0.1
MarkupSafe==1.1.0
oauthlib==3.0.0
Pillow==5.4.1
pkg-resources==0.0.0
python3-openid==3.1.0
pytz==2018.9
rcssmin==1.0.6
requests==2.21.0
requests-oauthlib==1.2.0
rjsmin==1.0.12
six==1.12.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.24.1
